Moving statistics to the statistics package.
